CAUSEWAY-3571: dont't expose resolver helpers; allows for further refactoring
