+++ commons/src/main/java/org/apache/causeway/commons/internal/reflection/_Reflect.java
@@ -40,7 +40,6 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.springframework.core.ResolvableType;
 import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.lang.Nullable;
 import org.springframework.util.ClassUtils;
@@ -52,7 +51,6 @@ import org.apache.causeway.commons.internal.assertions._Assert;
 import org.apache.causeway.commons.internal.base._NullSafe;
 import org.apache.causeway.commons.internal.base._Strings;
 import org.apache.causeway.commons.internal.collections._Arrays;
-import org.apache.causeway.commons.internal.context._Context;
 import org.apache.causeway.commons.internal.functions._Predicates;
 
 import lombok.NonNull;
@@ -616,81 +614,6 @@ public final class _Reflect {
 
     public Can<Constructor<?>> getPublicConstructors(final Class<?> cls) {
         return Can.ofArray(cls.getConstructors());
-    }
-
-    // -- GENENERIC TYPE ARG INTROSPECTION
-
-    //TODO[CAUSEWAY-3571] perhaps merge with ResolvedMethod
-    @lombok.Value(staticConstructor = "of")
-    public class MethodAndImplementingClass {
-        final @NonNull Method method;
-        final @NonNull Class<?> implementingClass;
-        /**
-         * [CAUSEWAY-3164] ensures reflection on generic type arguments works in a concurrent introspection setting
-         */
-        public Try<MethodAndImplementingClass> adopt(final @NonNull ClassLoader classLoader) {
-            try {
-                val ownerReloaded = Class.forName(implementingClass.getName(), true, classLoader);
-                val methodReloaded = ownerReloaded.getMethod(method.getName(), method.getParameterTypes());
-                return Try.success(MethodAndImplementingClass.of(methodReloaded, ownerReloaded));
-            } catch (Throwable e) {
-                return Try.failure(e);
-            }
-        }
-        public Try<MethodAndImplementingClass> adoptIntoDefaultClassLoader() {
-            return adopt(_Context.getDefaultClassLoader());
-        }
-        public Class<?> resolveFirstGenericTypeArgumentOnMethodReturn() {
-            return genericTypeArg(ResolvableType.forMethodReturnType(method, implementingClass))
-                    .toClass();
-        }
-        public Class<?> resolveFirstGenericTypeArgumentOnParameter(final int paramIndex) {
-            return genericTypeArg(ResolvableType.forMethodParameter(method, paramIndex, implementingClass))
-                    .toClass();
-        }
-        // -- HELPER
-        private static ResolvableType genericTypeArg(final ResolvableType nonScalar){
-            val genericTypeArg = nonScalar.isArray()
-                    ? nonScalar.getComponentType()
-                    : nonScalar.getGeneric(0);
-            return genericTypeArg;
-        }
-    }
-
-    @lombok.Value(staticConstructor = "of")
-    public class ConstructorAndImplementingClass {
-        final @NonNull Constructor<?> constructor;
-        final @NonNull Class<?> implementingClass;
-        /**
-         * [CAUSEWAY-3164] ensures reflection on generic type arguments works in a concurrent introspection setting
-         */
-        public Try<ConstructorAndImplementingClass> adopt(final @NonNull ClassLoader classLoader) {
-            try {
-                val ownerReloaded = Class.forName(implementingClass.getName(), true, classLoader);
-                val methodReloaded = ownerReloaded.getConstructor(constructor.getParameterTypes());
-                return Try.success(ConstructorAndImplementingClass.of(methodReloaded, ownerReloaded));
-            } catch (Throwable e) {
-                return Try.failure(e);
-            }
-        }
-        public Try<ConstructorAndImplementingClass> adoptIntoDefaultClassLoader() {
-            return adopt(_Context.getDefaultClassLoader());
-        }
-//        public Class<?> resolveFirstGenericTypeArgumentOnMethodReturn() {
-//            return genericTypeArg(ResolvableType.forConstructorParameter(executable, implementingClass))
-//                    .toClass();
-//        }
-        public Class<?> resolveFirstGenericTypeArgumentOnParameter(final int paramIndex) {
-            return genericTypeArg(ResolvableType.forConstructorParameter(constructor, paramIndex, implementingClass))
-                    .toClass();
-        }
-        // -- HELPER
-        private static ResolvableType genericTypeArg(final ResolvableType nonScalar){
-            val genericTypeArg = nonScalar.isArray()
-                    ? nonScalar.getComponentType()
-                    : nonScalar.getGeneric(0);
-            return genericTypeArg;
-        }
     }
 
     // -- FILTER
